Migrate another rust codegen test
